Move XML loading code to AnimationSet
This requires adding a new constructor taking just a fileName, and using
a ResourceCache for the loaded data.
